% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type3_bf.R
\name{bf_table}
\alias{bf_table}
\title{Bayes Factor Table}
\usage{
bf_table(x)
}
\arguments{
\item{x}{A list of BayesFactor objects.}
}
\value{
A tibble with the tested term, the Bayes Factor, and the error.
}
\description{
This function generates a table of Bayes Factors for a given list of
BayesFactor objects.
}
