% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type3_bf.R
\name{list_type3_models}
\alias{list_type3_models}
\title{List Type 3 Models}
\usage{
list_type3_models(x, dv, random)
}
\arguments{
\item{x}{A formula object representing the ful model.}

\item{dv}{A character string to identify the dependent variable. Not used
in the function, kept for compatibility.}

\item{random}{A character string to identify random terms in the model.}
}
\value{
A vecotr of all model formulas required for Type-3 model
compariosons. The full model is always the last vector element.
}
\description{
This function generates a list of submodels for a given model formula to
perform Type-3 mode comparisons, i.e. dropping individual terms from the
full model.
}
\examples{
\dontrun{
list_type3_models(y ~ x1 + x2 + id, dv = "y", random = "id")
}
}
