% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mochasta1}
\alias{mochasta1}
\alias{mochasta2}
\title{Experiment 1}
\format{
A data.frame with 7:
\describe{
\item{id}{Participant ID}
\item{task}{Task modality (spatial or verbal)}
\item{sound}{Sound condition (quiet, steady state, or variable state)}
\item{total}{}
\item{order}{Task order (Spatial or Verbal)}
\item{position}{Serial position (sp1 - sp7)}
\item{pos_total}{}
}

A data.frame with 6 variables:
\describe{
\item{id}{Participant ID}
\item{task}{Task modality (spatial or verbal)}
\item{sound}{Sound condition (quiet, steady state, or variable state)}
\item{total}{}
\item{position}{Serial position (sp1 - sp7)}
\item{pos_total}{}
}
}
\usage{
mochasta1

mochasta2
}
\description{
Data from Experiment 1

Data from Experiment 2
}
\keyword{data}
