% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type3_bf.R
\name{tested_terms}
\alias{tested_terms}
\title{Tested Terms}
\usage{
tested_terms(x)
}
\arguments{
\item{x}{A BayesFactor object.}
}
\value{
A list of terms that were tested.
}
\description{
This function extracts the terms that were tested in a BayesFactor object.
}
